test: merge duplicated node search and asset browser Playwright families
Keep a union of unique behaviors: V1 node search stays as a still-shipping
UI variant, V2 specs fold into one file, and asset-browser coverage moves
under tests/assets/ by behavior. Drop only true duplicates (weaker V2
category filter, shadow-backend display/footer copies, and the fixme
media-type filter already covered by the cloud filter spec).
