+++ browser_tests/tests/nodeSearch/nodeSearchBox.spec.ts-snapshots/added-node-chromium-linux.png
