+++ browser_tests/tests/nodeSearch/nodeSearchBox.spec.ts-snapshots/added-node-no-connection-chromium-linux.png
