+++ browser_tests/tests/nodeSearch/nodeSearchBox.spec.ts-snapshots/auto-linked-node-batch-chromium-linux.png
