+++ browser_tests/tests/nodeSearch/nodeSearchBox.spec.ts-snapshots/auto-linked-node-chromium-linux.png
