+++ browser_tests/tests/nodeSearch/nodeSearchBox.spec.ts-snapshots/link-context-menu-search-chromium-linux.png
