+++ browser_tests/tests/nodeSearch/nodeSearchBox.spec.ts-snapshots/link-release-context-menu-chromium-linux.png
